fix:別所 Task.25 削除フラグを定数を使用するよう修正。
StudentAttendanceService:382行目
TStudentAttendanceMapper:78行目

+++ trial_source/src/main/java/jp/co/sss/lms/mapper/TStudentAttendanceMapper.java
@@ -75,6 +75,6 @@ public interface TStudentAttendanceMapper {
 	 * @return 未入力件数
 	 */
 	Integer notEnterCount(@Param("lmsUserId") Integer lmsUserId,
-			@Param("deleteFlg") Integer deleteFlg,
+			@Param("deleteFlg") short deleteFlg,
 			@Param("trainingDate") String trainingDate);
 }
